Evaluate matrix arithmetic and record outputs in the DAE interpreter
- Evaluate elementwise and scalar-broadcast matrix binaries, record-field
  matrix operands, and rank-2 shape-defining builtins as true matrix
  values, so chains like 'A + 0.5*A*B + Om*A*(C1*I + C2*B)' follow
  MLS 10.6 matrix-product semantics instead of failing elementwise
- Treat self-referential input shapes ('A[:, size(A, 1)]') as
  argument-inferred dynamic dims
- Materialize record-typed outputs in multi-output call statements
  ('(corrected, accepted) := correct(...)') by copying output fields
  and dims to the target record
- Split matrix evaluation into eval/array_eval/matrix_eval.rs to meet
  the SPEC_0021 file-size budget

Tests.EstimationTests (full IEKF prediction/correction suite) now
interprets natively; PlanningTests and the CI subset stay green.
